Fix configuration insertion with Thinking Sphinx v4.2.
The new config preparation callback in Thinking Sphinx's Railtie was overwriting the saved settings from Flying Sphinx, and thus the non-localhost connection and specific port were being forgotten. So, now we've got the same callback defined here in Flying Sphinx. The order will be consistent of TS before FS, because FS loads TS as a dependency (thus, the order in an app's Gemfile is irrelevant).